Add configuration options for property execution
Add configuration option to control the number of times each property is
executed. The number can be specified in the rebar.config file as an
entry under eqc_opts or overridden on the command line. Also provide and
option for specifying a time duration for each property to execute
instead of a fixed number of iterations. This can also be set in the
rebar.config file or overridden on the command line.